+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - display the next four clock changes for configured time values with DST in info
 
+### Changed
+- a few common timezone abbreviations (MEZ, MESZ, OEZ, OESZ, WEZ, WESZ) localized for German
+
 ---
 
 ## [1.3.0] - 2026-07-20